requirements.txt
django
psycopg2_binary
djangorestframework
